Fix errors

+++ kernel/include/pmm.h
@@ -9,7 +9,7 @@ uint8_t* free_mem;
 void pmm_set_used(void *, void *);
 void pmm_init(void *);
 void pmm_free(void *);
-void *pmm_alloc(int);
+void *pmm_alloc(size_t);
 void pmm_mark_used(int);
 
 #endif
